dNdS

dN / dS

NaturalSelection.jl provides several different methods of inferring the action of natural selection from coding sequences.

Evolutionary pressures on proteins are often quantified by the ratio of substitution rates at non-synonymous and synonymous sites i.e. dN/dS.

The dN/dS ratio was originally developed for application to distantly diverged sequences, the differences among which represent substitutions that have fixed along independent lineages.

Nevertheless, the dN/dS measure is often applied to sequences sampled from a single population, the differences among which represent segregating polymorphisms. However, do be careful if this is what you are doing, as it has been demonstrated that dN/dS is not always suitable for such purposes (Sergey Kryazhimskiy & Joshua B. Plotkin, 2008).

The NG86 method

NaturalSelection.dNdS_NG86 — Function.

dNdS_NG86(x, y, k::Float64 = 1.0, code::GeneticCode)

Compute dN and dS, using the Nei and Goborjei 1986 method.

This function requires two iterables x and y, which yield DNACodon or RNACodon type variables. These two types are defined in the BioSequences package.

source

dNdS_NG86(x::BioSequence{A}, y::BioSequence{A}, k::Float64, code::GeneticCode) where {A <: NucAlphs}

Compute dN and dS, using the Nei and Goborjei 1986 method.

This method adds conveinience when working with DNA or RNA sequences, by taking two sequences, and creating two vectors of aligned codons from them. These two iterables are then passed into the generic NG86 method.

source

NaturalSelection.S_N_NG86 — Function.

S_N_NG86(codon::C, k::Float64, code::GeneticCode) where {C <: CDN}

Enumerate the number of synonymous (S) and non-synonymous (N) sites in a codon, using the method used by Nei and Goborjei (1986).

Returns a tuple where S is the first element and N is the second (S, N).

Each site in a codon may be both partially synonymous and non-synonymous.

source

NaturalSelection.DS_DN_NG86 — Function.

DS_DN_NG86(x::C, y::C, code::GeneticCode) where C <: CDN

Compute the number of synonymous (DS) and non-synonymous (DN) mutations between two codons, using the all paths method used by the Nei and Goborjei (1986).

source